% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{Account}
\alias{Account}
\title{Account Class}
\description{
This base class for account objects.
}
\details{
An Account has a balance.
One can deposit or withdraw some money from an account.
}
\seealso{
Other Account: 
\code{\link{GiroAccount}},
\code{\link{SafeAccount}}
}
\concept{Account}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{balance}}{Account balance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-deposit}{\code{Account$deposit()}}
\item \href{#method-withdraw}{\code{Account$withdraw()}}
\item \href{#method-clone}{\code{Account$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit"></a>}}
\if{latex}{\out{\hypertarget{method-deposit}{}}}
\subsection{Method \code{deposit()}}{
deposit a certain amount (>0) into the account
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$deposit(amount = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{(\code{integer()} >0) amount be deposited.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the object itself, but modified \strong{by reference}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\if{latex}{\out{\hypertarget{method-withdraw}{}}}
\subsection{Method \code{withdraw()}}{
withdraw a certain amount (>0) from the account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$withdraw(amount = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{(\code{integer()} >0) amount be withdrawn.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the object itself, but modified \strong{by reference}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
