% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.R
\name{Account_log}
\alias{Account_log}
\title{R6 class providing an account with a log}
\description{
R6 class providing an account with a log

R6 class providing an account with a log
}
\details{
An account log has a value and a transaction log.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Account_log balance.}

\item{\code{TransactionLog}}{transaction log.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Account_log$new()}}
\item \href{#method-print}{\code{Account_log$print()}}
\item \href{#method-deposit}{\code{Account_log$deposit()}}
\item \href{#method-withdraw}{\code{Account_log$withdraw()}}
\item \href{#method-clone}{\code{Account_log$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new bank account with a log.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account_log$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new "Account_log" object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print the value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account_log$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit"></a>}}
\subsection{Method \code{deposit()}}{
Deposit money.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account_log$deposit(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{How much money.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\subsection{Method \code{withdraw()}}{
Withdraw money.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account_log$withdraw(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{how much money.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account_log$clone(deep = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
