% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.R
\name{giro_account}
\alias{giro_account}
\title{R6 Subclass for a giro account}
\description{
A special form of bank account that has an overdraft limit and fee
}
\section{Super class}{
\code{\link[bankr:Account]{bankr::Account}} -> \code{GiroAccount}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-withdraw}{\code{giro_account$withdraw()}}
\item \href{#method-clone}{\code{giro_account$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="deposit">}\href{../../bankr/html/Account.html#method-deposit}{\code{bankr::Account$deposit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\if{latex}{\out{\hypertarget{method-withdraw}{}}}
\subsection{Method \code{withdraw()}}{
method to withdraw money from giro account
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{giro_account$withdraw(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{the amount of money to withdraw from account}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{giro_account$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
