% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiroAccount.R
\name{GiroAccount}
\alias{GiroAccount}
\title{Giro Account Class}
\description{
This Account extends \link{Account} with a withdrawal limit.
}
\seealso{
Other Account: 
\code{\link{Account}},
\code{\link{SafeAccount}}
}
\concept{Account}
\section{Super class}{
\code{\link[bankr:Account]{bankr::Account}} -> \code{GiroAccount}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{limit}}{(\code{integer()} >0) withdraw limit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GiroAccount$new()}}
\item \href{#method-withdraw}{\code{GiroAccount$withdraw()}}
\item \href{#method-clone}{\code{GiroAccount$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="deposit">}\href{../../bankr/html/Account.html#method-deposit}{\code{bankr::Account$deposit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$new(limit = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{(\code{integer()} >0) withdraw limit.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\if{latex}{\out{\hypertarget{method-withdraw}{}}}
\subsection{Method \code{withdraw()}}{
withdrawals above the limit are not possible.
withdrawals from an account with negative balance is penalized
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$withdraw(amount = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{(\code{integer()} >0) amount be withdrawn.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the object itself, but modified \strong{by reference}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
