% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wf.R
\name{read_wf}
\alias{read_wf}
\title{Read Wells Fargo Account Statement (Checking)}
\usage{
read_wf(file)
}
\arguments{
\item{file}{File path to the checking account statement PDF}
}
\value{
A tibble containing transactional data
}
\description{
Read Wells Fargo Account Statement (Checking)
}
\examples{
read_wf("Checking_Account.pdf")
}
