% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wf_checking.R
\name{read_wf_checking}
\alias{read_wf_checking}
\title{Read Wells Fargo Account Statement (Checking)}
\usage{
read_wf_checking(file, type = "detail")
}
\arguments{
\item{file}{File path to the checking account statement PDF}

\item{type}{Specifies output type. detail = transaction details. summary = statement summary}
}
\value{
A tibble containing transactional data
}
\description{
Read Wells Fargo Account Statement (Checking)
}
