% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wf_credit.R
\name{read_wf_credit}
\alias{read_wf_credit}
\title{Read Wells Fargo Statement - Credit}
\usage{
read_wf_credit(file, type = "detail")
}
\arguments{
\item{file}{File path to the credit account statement PDF}

\item{type}{Specifies output type. detail = transaction details. summary = statement summary}
}
\value{
A tibble containing transactional data
}
\description{
Read Wells Fargo Statement - Credit
}
\examples{

}
