% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.R
\name{safe_account}
\alias{safe_account}
\title{R6 Class for a safe bank account}
\description{
A bank account with a private field for balance instead of a public field
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{deposit}}{Deposits money to the safe account and returns the balance
after the transaction is finished.}

\item{\code{get_balance}}{Returns the current account balance.}

\item{\code{withdraw}}{Withdraws money from the safe account and returns the balance
after the transaction is finished.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{safe_account$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{safe_account$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
