% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account-r6.R
\name{SafeAccount}
\alias{SafeAccount}
\title{R6 Object representing a safe bank account}
\description{
Implements an R6 structure for safe bank accounts and corresponding actions
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{withdraw}}{Withdraw money from bank account or providing access current balance.
If missing argument then the currenct balance will be returned}

\item{\code{deposit}}{Deposit money into bank account or providing access current balance.
If missing argument then the currenct balance will be returned}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SafeAccount$new()}}
\item \href{#method-clone}{\code{SafeAccount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new SafeAccount object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafeAccount$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `SafeAccount` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafeAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
