% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.R
\name{TransactionLog}
\alias{TransactionLog}
\title{R6 Class representing a transaction log}
\description{
R6 Class representing a transaction log

R6 Class representing a transaction log
}
\details{
A transaction log stores a history of transactions for a bank account.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{history}}{Expose private$..history}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-append_transaction}{\code{TransactionLog$append_transaction()}}
\item \href{#method-print}{\code{TransactionLog$print()}}
\item \href{#method-clone}{\code{TransactionLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_transaction"></a>}}
\subsection{Method \code{append_transaction()}}{
Append a new transaction to current history.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionLog$append_transaction(type, amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{One of c("withdraw", "deposit", "transfer")}

\item{\code{amount}}{Amount of money processed by transaction}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print transaction history
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionLog$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
Dataframe with date, transaction and amount as colums
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
