% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account-r6.R
\name{Account}
\alias{Account}
\title{R6 Object representing a bank account}
\description{
Implements an R6 structure for bank accounts and corresponding actions
}
\examples{

## ------------------------------------------------
## Method `Account$withdraw`
## ------------------------------------------------

A <- Account$new()
A$deposit(100)
A$withdraw(50)

## ------------------------------------------------
## Method `Account$deposit`
## ------------------------------------------------

A <- Account$new()
A$deposit(100)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{balance}}{Current account balance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Account$new()}}
\item \href{#method-withdraw}{\code{Account$withdraw()}}
\item \href{#method-deposit}{\code{Account$deposit()}}
\item \href{#method-clone}{\code{Account$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new account object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `Account` object with `balance` of 0.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\subsection{Method \code{withdraw()}}{
Withdraw money from bank account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$withdraw(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Amount to be withdrawn. Must be a whole number.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A <- Account$new()
A$deposit(100)
A$withdraw(50)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit"></a>}}
\subsection{Method \code{deposit()}}{
Deposit money into bank account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$deposit(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Amount to be deposited. Must be a whole number.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{A <- Account$new()
A$deposit(100)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
