% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.R
\name{Account}
\alias{Account}
\title{R6 Class representing a general bank account}
\description{
R6 Class representing a general bank account

R6 Class representing a general bank account
}
\details{
An account has a balance and provides operations on this balance.
An account also has a history of transactions @seealso [TransactionLog].
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{balance}}{Expose private$..balance}

\item{\code{transaction_log}}{Expose methods to read and write private$..t_log}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Account$new()}}
\item \href{#method-validate}{\code{Account$validate()}}
\item \href{#method-deposit}{\code{Account$deposit()}}
\item \href{#method-withdraw}{\code{Account$withdraw()}}
\item \href{#method-clone}{\code{Account$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new Account object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$new(balance = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{balance}}{Initial account balance.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Account` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\subsection{Method \code{validate()}}{
Validate a freshly opened bank account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit"></a>}}
\subsection{Method \code{deposit()}}{
Deposit money into account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$deposit(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Amount of money}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\subsection{Method \code{withdraw()}}{
Withdraw money from account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$withdraw(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Amount of money}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$clone()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
