% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.R
\name{Account}
\alias{Account}
\title{R6 class providing an account}
\description{
R6 class providing an account

R6 class providing an account
}
\details{
An account has a value.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Account balance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Account$new()}}
\item \href{#method-print}{\code{Account$print()}}
\item \href{#method-deposit}{\code{Account$deposit()}}
\item \href{#method-withdraw}{\code{Account$withdraw()}}
\item \href{#method-clone}{\code{Account$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new bank account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new "Account" object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print the value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit"></a>}}
\subsection{Method \code{deposit()}}{
Deposit money.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$deposit(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{How much money.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\subsection{Method \code{withdraw()}}{
Withdraw money.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$withdraw(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{how much money.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
