% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{Account}
\alias{Account}
\title{Account class}
\description{
R6 class to represent a general bank account

The bank account comes with the following basic functionality and has
the private fields \code{..balance} and \code{..account_holder} that store the most
important information.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{balance}}{read only access to variable \code{balance}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Account$new()}}
\item \href{#method-deposit}{\code{Account$deposit()}}
\item \href{#method-withdraw}{\code{Account$withdraw()}}
\item \href{#method-print}{\code{Account$print()}}
\item \href{#method-clone}{\code{Account$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize bank account and check the inputs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$new(account_holder, balance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_holder}}{single character vector}

\item{\code{balance}}{single numeric value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deposit"></a>}}
\if{latex}{\out{\hypertarget{method-deposit}{}}}
\subsection{Method \code{deposit()}}{
deposit money on the bank account
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$deposit(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{single non negative value to deposit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\if{latex}{\out{\hypertarget{method-withdraw}{}}}
\subsection{Method \code{withdraw()}}{
withdraw money from the bank account
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$withdraw(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{single non negative value to withdraw}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print account holder and balance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
just prints in the console
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Account$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
