% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account-r6.R
\name{GiroAccount}
\alias{GiroAccount}
\title{R6 Object representing a giro bank account}
\description{
Implements an R6 structure for giro bank accounts and corresponding actions
}
\section{Super class}{
\code{\link[bankr:Account]{bankr::Account}} -> \code{GiroAccount}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{limit}}{Overdraft limit.}

\item{\code{overdraft_fee}}{Overdraft fee.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GiroAccount$new()}}
\item \href{#method-withdraw}{\code{GiroAccount$withdraw()}}
\item \href{#method-clone}{\code{GiroAccount$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="deposit">}\href{../../bankr/html/Account.html#method-deposit}{\code{bankr::Account$deposit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new GiroAccount object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$new(limit = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{Overdraft limit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `GiroAccount` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\subsection{Method \code{withdraw()}}{
Withdraw money from bank account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$withdraw(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Amount to be withdrawn. Must be a whole number.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If more money is withdrawn than client has the exceeding amount will be charged with
overdraft fees. If the amount withdrawn would exceed the overdraft limit, no withdrawal will be made.
is possible and an error message is returned.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
