% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.R
\name{GiroAccount}
\alias{GiroAccount}
\title{R6 Class representing a giro bank account}
\description{
R6 Class representing a giro bank account

R6 Class representing a giro bank account
}
\details{
It has an overdraft limit which generates a small fee when exceeded.
}
\seealso{
[Account]
}
\section{Super class}{
\code{\link[bankr:Account]{bankr::Account}} -> \code{GiroAccount}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GiroAccount$new()}}
\item \href{#method-validate}{\code{GiroAccount$validate()}}
\item \href{#method-withdraw}{\code{GiroAccount$withdraw()}}
\item \href{#method-clone}{\code{GiroAccount$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="deposit">}\href{../../bankr/html/Account.html#method-deposit}{\code{bankr::Account$deposit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new GiroAccount object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$new(balance = 0, overdraft_limit = Inf, overdraft_fee = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{balance}}{Initial account balance.}

\item{\code{overdraft_limit}}{Nonchargeable limit}

\item{\code{overdraft_fee}}{Chargable fee when `overdraft_limit` exceeded}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `GiroAccount` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\subsection{Method \code{validate()}}{
Validate a freshly opened bank account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\subsection{Method \code{withdraw()}}{
Draw additional fee when overdraft limit exceeded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$withdraw(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{Amount of money}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
