% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.R
\name{GiroAccount}
\alias{GiroAccount}
\title{R6 class providing a giro account}
\description{
R6 class providing a giro account

R6 class providing a giro account
}
\details{
A giro account has a value and a limit.
}
\section{Super class}{
\code{\link[bankr:Account]{bankr::Account}} -> \code{GiroAccount}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{limit}}{GiroAccount limit.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GiroAccount$new()}}
\item \href{#method-withdraw}{\code{GiroAccount$withdraw()}}
\item \href{#method-clone}{\code{GiroAccount$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="deposit">}\href{../../bankr/html/Account.html#method-deposit}{\code{bankr::Account$deposit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="print">}\href{../../bankr/html/Account.html#method-print}{\code{bankr::Account$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new giro account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$new(limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{overdraw limit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new "GiroAccount" object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\subsection{Method \code{withdraw()}}{
Withdraw money.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$withdraw(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{how much money.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
