% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giroAccount.R
\name{GiroAccount}
\alias{GiroAccount}
\title{GiroAccount class}
\description{
R6 class to represent a Giro bank account

Additional features to the general \code{Account} are a maximal overdraft
limit as well as a fee to punish negative account balances.
}
\section{Super class}{
\code{\link[bankr:Account]{bankr::Account}} -> \code{GiroAccount}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GiroAccount$new()}}
\item \href{#method-withdraw}{\code{GiroAccount$withdraw()}}
\item \href{#method-clone}{\code{GiroAccount$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="deposit">}\href{../../bankr/html/Account.html#method-deposit}{\code{bankr::Account$deposit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="print">}\href{../../bankr/html/Account.html#method-print}{\code{bankr::Account$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize bank account and check the inputs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$new(
  account_holder,
  balance,
  overdraft_max = Inf,
  overdraft_fee_rate = 0.09
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_holder}}{single character vector}

\item{\code{balance}}{single numeric value}

\item{\code{overdraft_max}}{single non negative numeric value}

\item{\code{overdraft_fee_rate}}{single numeric value between 0 and 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\if{latex}{\out{\hypertarget{method-withdraw}{}}}
\subsection{Method \code{withdraw()}}{
withdraw money from the giro account.
charges fee if giro account negative. throws error,
if overdraft limit is reached
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$withdraw(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{single non negative value to withdraw}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
