% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bankkonto-sol.R
\name{GiroAccount}
\alias{GiroAccount}
\title{Giro Account Class}
\description{
Class for a giro bank account which enables the deposit and withdraw of money within a certain overdraft limit
}
\section{Super class}{
\code{\link[bankr:Account]{bankr::Account}} -> \code{GiroAccount}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{account_holder}}{name of the person owning the bank account}

\item{\code{balance}}{current available credit on bank account}

\item{\code{overdraft_max}}{limit of negative credit the user is allowed to have}

\item{\code{overdraft_fee_rate}}{each time a withdrawing results in a negative credit, 'overdraft_fee_rate' times the resulting credit is additionally subtracted from the 'balance'}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{account_holder}}{name of the person owning the bank account}

\item{\code{balance}}{current available credit on bank account}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GiroAccount$new()}}
\item \href{#method-withdraw}{\code{GiroAccount$withdraw()}}
\item \href{#method-clone}{\code{GiroAccount$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="deposit">}\href{../../bankr/html/Account.html#method-deposit}{\code{bankr::Account$deposit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bankr" data-topic="Account" data-id="print">}\href{../../bankr/html/Account.html#method-print}{\code{bankr::Account$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$new(account_holder, balance, overdraft_max, overdraft_fee_rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_holder}}{name of the person owning the bank account}

\item{\code{balance}}{starting credit of the bank account}

\item{\code{overdraft_max}}{limit of negative credit the user is allowed to have}

\item{\code{overdraft_fee_rate}}{each time a withdrawing results in a negative credit, 'overdraft_fee_rate' times the resulting credit is additionally subtracted from the 'balance'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-withdraw"></a>}}
\if{latex}{\out{\hypertarget{method-withdraw}{}}}
\subsection{Method \code{withdraw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$withdraw(amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{amount}}{amount of money which should be transferred}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GiroAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
