% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankr.R
\name{SafeAccount}
\alias{SafeAccount}
\title{R6 class providing a safe account}
\description{
R6 class providing a safe account

R6 class providing a safe account
}
\details{
A safe account has a value.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{deposit}}{active binding to value.}

\item{\code{withdraw}}{active binding to value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SafeAccount$new()}}
\item \href{#method-print}{\code{SafeAccount$print()}}
\item \href{#method-clone}{\code{SafeAccount$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new bank account.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafeAccount$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new "SafeAccount" object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print the value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafeAccount$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SafeAccount$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
