% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account-r6.R
\name{TransactionLog}
\alias{TransactionLog}
\title{R6 Object representing a transaction log}
\description{
Implements an R6 structure for a transaction log and corresponding actions
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{trans_log}}{Provides access to current transition log.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-make_trans}{\code{TransactionLog$make_trans()}}
\item \href{#method-clone}{\code{TransactionLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_trans"></a>}}
\subsection{Method \code{make_trans()}}{
Make a transaction and add to the transaction log.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionLog$make_trans(trans_type, amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trans_type}}{The transaction type. Must be either 'deposit' or 'withdraw'.}

\item{\code{amount}}{Amount to be deposited. Must be a whole number.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with the columns 'Date', 'Transaction', 'Amount', 'Balance'.
Each row describes a transaction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TransactionLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
